Remove SSHFormPage and associated files
This commit removes the `SSHFormPage.xaml` and its code-behind file `SSHFormPage.xaml.cs` from the project, indicating that this page is no longer part of the application. Additionally, the entry for `SSHFormPage.xaml` has been removed from the project file `BarkodListem.csproj`. The `SSHFormViewModel.cs` file has also been modified to eliminate the class definition for `SSHFormViewModel`, which was associated with the SSH form.